modified test case
