+++ calyx-opt/src/passes/mod.rs
@@ -34,7 +34,6 @@ mod static_promotion;
 mod sync;
 // mod simplify_guards;
 mod add_guard;
-mod compile_static_interface;
 mod data_path_infer;
 mod default_assigns;
 mod discover_external;
@@ -84,7 +83,6 @@ pub use sync::CompileSync;
 pub use sync::CompileSyncWithoutSyncReg;
 // pub use simplify_guards::SimplifyGuards;
 pub use add_guard::AddGuard;
-pub use compile_static_interface::CompileStaticInterface;
 pub use default_assigns::DefaultAssigns;
 pub use synthesis_papercut::SynthesisPapercut;
 pub use top_down_compile_control::TopDownCompileControl;
