+++ tests/passes/compile-static-interface/compile-static-interface-one-cycle.expect
@@ -6,8 +6,6 @@ static<1> component do_add(left: 32, right: 32, @go go: 1, @clk clk: 1, @reset r
     r = std_reg(32);
   }
   wires {
-    static<1> group a {
-    }
     r.write_en = go ? 1'd1;
     add.right = go ? right;
     add.left = go ? left;
@@ -18,15 +16,33 @@ static<1> component do_add(left: 32, right: 32, @go go: 1, @clk clk: 1, @reset r
 component main(@go go: 1, @clk clk: 1, @reset reset: 1) -> (@done done: 1) {
   cells {
     a = do_add();
+    @generated fsm = std_reg(1);
+    @generated adder = std_add(1);
+    @generated ud = undef(1);
+    @generated signal_reg = std_reg(1);
   }
   wires {
-    static<1> group static_invoke {
-      a.go = %0 ? 1'd1;
+    group early_reset_static_invoke {
+      a.go = fsm.out == 1'd0 ? 1'd1;
       a.left = 32'd5;
       a.right = 32'd6;
+      adder.left = fsm.out;
+      adder.right = 1'd1;
+      fsm.write_en = 1'd1;
+      fsm.in = fsm.out != 1'd0 ? adder.out;
+      fsm.in = fsm.out == 1'd0 ? 1'd0;
+      early_reset_static_invoke[done] = ud.out;
     }
+    group wrapper_early_reset_static_invoke {
+      early_reset_static_invoke[go] = 1'd1;
+      signal_reg.write_en = fsm.out == 1'd0 & !signal_reg.out ? 1'd1;
+      signal_reg.in = fsm.out == 1'd0 & !signal_reg.out ? 1'd1;
+      wrapper_early_reset_static_invoke[done] = fsm.out == 1'd0 & signal_reg.out ? 1'd1;
+    }
+    signal_reg.write_en = fsm.out == 1'd0 & signal_reg.out ? 1'd1;
+    signal_reg.in = fsm.out == 1'd0 & signal_reg.out ? 1'd0;
   }
   control {
-    static_invoke;
+    wrapper_early_reset_static_invoke;
   }
 }
