+++ tests/passes/compile-static-interface/compile-static-interface-one-cycle.futil
@@ -1,4 +1,4 @@
-// -p validate -p compile-invoke -p static-inline -p add-guard -p compile-static-interface
+// -p validate -p compile-invoke -p static-inline -p add-guard -p compile-static
 import "primitives/core.futil";
 import "primitives/memories/comb.futil";
