+++ tests/passes/compile-static/interface-and-cs.expect
@@ -6,16 +6,9 @@ static<6> component do_add(left: 32, right: 32, @go go: 1, @clk clk: 1, @reset r
     r = std_reg(32);
     @generated fsm = std_reg(3);
     @generated adder = std_add(3);
-    @generated fsm0 = std_reg(3);
-    @generated adder0 = std_add(3);
-    @generated adder1 = std_add(3);
-    @generated adder2 = std_add(3);
+    @generated fsm0 = std_reg(1);
+    @generated adder0 = std_add(1);
     @generated ud = undef(1);
-    @generated ud0 = undef(1);
-    @generated ud1 = undef(1);
-    @generated fsm1 = std_reg(3);
-    @generated adder3 = std_add(3);
-    @generated ud2 = undef(1);
   }
   wires {
     group early_reset_a {
@@ -24,10 +17,10 @@ static<6> component do_add(left: 32, right: 32, @go go: 1, @clk clk: 1, @reset r
       add.left = r.out;
       r.in = add.out;
       adder0.left = fsm0.out;
-      adder0.right = 3'd1;
+      adder0.right = 1'd1;
       fsm0.write_en = 1'd1;
-      fsm0.in = fsm0.out != 3'd0 ? adder0.out;
-      fsm0.in = fsm0.out == 3'd0 ? 3'd0;
+      fsm0.in = fsm0.out != 1'd0 ? adder0.out;
+      fsm0.in = fsm0.out == 1'd0 ? 1'd0;
       early_reset_a[done] = ud.out;
     }
     r.write_en = go & fsm.out == 3'd0 ? 1'd1;
